Everything is ok, some refactoring is done
