+++ src/main.js
@@ -12,7 +12,7 @@ import EditingPoint from './view/editing-point.js';
 // import NewWithoutOffers from './view/new-without-offers.js';
 // import NewPoint from './view/newPoint.js';
 import { generatePoint } from './mock/create-data.js';
-import { renderPosition, render } from './view/rendering-utils.js';
+import { renderPosition, render, replace } from './utils/rendering-utils.js';
 
 const COUNT_OF_POINTS = 15;
 
@@ -24,31 +24,31 @@ const toNavigation = document.querySelector('.trip-controls__navigation');
 const toFilters = document.querySelector('.trip-controls__filters');
 const toSort = document.querySelector('.trip-events');
 
-render(priceAndTripSection, new PriceTripView(points).getElement(), renderPosition.AFTERBEGIN);
-render(toNavigation, new NavigationView().getElement(), renderPosition.AFTERBEGIN);
-render(toFilters, new FiltersView().getElement(), renderPosition.AFTERBEGIN);
+render(priceAndTripSection, new PriceTripView(points), renderPosition.AFTERBEGIN);
+render(toNavigation, new NavigationView(), renderPosition.AFTERBEGIN);
+render(toFilters, new FiltersView(), renderPosition.AFTERBEGIN);
 
 if (points.length === 0) {
-  render(toSort, new Empty().getElement(), renderPosition.BEFOREEND);
+  render(toSort, new Empty(), renderPosition.BEFOREEND);
 } else {
-  render(toSort, new SortMenuView().getElement(), renderPosition.AFTERBEGIN);
+  render(toSort, new SortMenuView(), renderPosition.AFTERBEGIN);
 }
 
 const createUl = new TripListUl();
-render(toSort, createUl.getElement(), renderPosition.BEFOREEND);
+render(toSort, createUl, renderPosition.BEFOREEND);
 
 const renderPoint = (point) => {
   const pointEvent = new PointsList(point);
   const editPoint = new EditingPoint(point);
   const tripListLi = new TripListLi();
-  render(createUl.getElement(), tripListLi.getElement(), renderPosition.BEFOREEND);
+  render(createUl, tripListLi, renderPosition.BEFOREEND);
 
   const replaceCardToForm = () => {
-    tripListLi.getElement().replaceChild(editPoint.getElement(), pointEvent.getElement());
+    replace(editPoint, pointEvent);
   };
 
   const replaceFormToCard = () => {
-    tripListLi.getElement().replaceChild(pointEvent.getElement(), editPoint.getElement());
+    replace(pointEvent, editPoint);
   };
 
   const onEscKeyDown = (evt) => {
@@ -74,12 +74,10 @@ const renderPoint = (point) => {
     document.removeEventListener('keydown', onEscKeyDown);
   });
 
-  render(tripListLi.getElement(), pointEvent.getElement(), renderPosition.BEFOREEND);
+  render(tripListLi, pointEvent, renderPosition.BEFOREEND);
 };
 
-for (const point of points) {
-  renderPoint(point);
-}
+points.forEach((point) => renderPoint(point));
 
 // render(toSort, new Loading().getElement(), renderPosition.BEFOREEND);
 // render(toSort, new Empty().getElement(), renderPosition.BEFOREEND);
