+++ src/view/editing-point.js
@@ -1,5 +1,5 @@
 import { TYPES, CITIES } from '../mock/create-data.js';
-import { addOffers, createTypes, createCities, getFormatTime } from '../utils/rendering-utils.js';
+import { addOffers, createTypes, createCities, getFormatTime } from '../utils/rendering-data-utils.js';
 import Abstract from '../view/abstract.js';
 
 const editPoint = (points) => {
